Clean up admin data fetching
