+++ apps/web/src/components/product/QuantityControl.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Minus, Plus } from 'lucide-react';
 import { Button } from '../ui/button';
